requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.8
packaging==24.1
# psycopg2==2.9.9
psycopg2-binary
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.2
